Fix what a blind review of the LLVM 20 port found
Eight reviewers, one partition each, given the diff, the 20.1.8 headers and
the llvm-project source at both tags, and nothing anyone had said about the
branch. What they found that an audit of suspicions had not was contract
changes: upstream methods the shim still calls by the same name, whose
bodies changed underneath it. Every crash below was reproduced by probe
before it was touched.

Regressions from the port, fixed:

- `FieldDecl::getBitWidthValue` stopped evaluating the width and reads an
  integer cached in a `ConstantExpr` through an unchecked cast. A plain field
  segfaulted and a width installed with `setBitWidth` came back as garbage.
  It and `isZeroLengthBitField` now restate the precondition.
- `NonTypeTemplateParmDecl::getDefaultArgument` returns an empty
  `TemplateArgumentLoc` where it returned a null `Expr*`, and reading the
  expression out of that asserts: a parameter with no default aborted the
  process. The wrapper answers the documented null itself, and the C header
  no longer promises one.
- `Sema::getMoreSpecializedTemplate` gained `RawObj1Ty`/`RawObj2Ty`, which
  the shim filled with null types that clang dereferences; ordering two
  member function templates segfaulted where LLVM 18 worked. The shim and the
  wrapper take upstream's signature now. The object types default to the
  class each template belongs to, as overload resolution passes them, and a
  null one is refused exactly where clang would read it.
- Code generation reads `CompilerInstance::getTimerGroup()` under
  `-ftime-report`, which only `cc1_main` creates; the flag segfaulted
  `create_interpreter` and `create_irgenerator`. `CodeGenOptions` gains
  `getTimePasses`/`setTimePasses`, the drivers create the timer when it is
  set, and `EmitBackendOutput` restates the precondition.
- `Language::CIR` joined the languages `getDefaultLanguageStandard` answers
  with `llvm_unreachable`; two gates still listed only the other two.
- Ten enum mirrors lacked enumerators LLVM 20 appended, so `getCallConv` on
  a `preserve_none` function was `<invalid #22>`. ENUM_SYNC cannot see this
  -- nothing names the missing enumerator -- so each of the ten also gets a
  `switch` over the upstream enum with no default and `-Wswitch` as an error:
  the next appended enumerator stops the build and names itself.
  `CXTemplateName_NameKind` had no ENUM_SYNC table and one enumerator spelled
  without its prefix; both fixed.
- `Decl::getGlobalID` is 64 bits in LLVM 20, the module-file index in the
  upper half, and crossed as `unsigned`. It and the 47 `CreateDeserialized`
  ID parameters are `uint64_t`.
- `getMinGlobalAlign` hard-coded `false` for upstream's new `HasNonWeakDef`,
  the opposite of what clang passes when it has no declaration; it is a
  parameter, defaulting to `true`.

Reachable on main as well, fixed in passing: `CompileDtorCall` on a record
with no definition; a builtin ID past the end of the table, which
`Sema.CreateBuiltin` and fourteen `Builtin::Context` accessors indexed
unchecked -- `getNumBuiltins` reads the two private target tables' sizes so
the gate has an upper bound; `create_interpreter` returning a null
interpreter where construction failed, now an error that names
`create_parser`.

Also: a JLL with no build for the running LLVM is reported as that, rather
than as an `UndefVarError` from inside the JLL. The fabricated
`ModuleConstructorTag` is static_asserted empty; `getMaxBuiltinID` derives
from the sentinel instead of spelling 65533. The generator pins
`LLVM_full_jll` to the LLVM it names its output after: unpinned it installed
23.1.1, and handed a `VersionNumber` Pkg 1.13 fails to set it up, so the pin
is a string. Regenerating against 20.1.8 changed nothing but what the headers
above changed.

Tests. The critics of the test rewrite found what it still could not see:
fourteen more wrapper methods with no caller (a name-level check had missed
per-receiver overloads such as `getSourceRange`); seven `test/types.jl`
assertions turned from `isa Carrier` into `==`, which on carriers is pointer
identity and cannot fail; some thirty-five asm assertions behind guards with
no else; twenty-two `x === nothing || ...`, which pass whichever outcome
clang chose; an MSGuid fixture whose value existed only under LLP64; and
`# shape-only` markers whose reason was false, among them a module's
availability, which its constructor decides. All are pins, relations or both
polarities now, each checked under the x86_64 Linux and Windows triples by
cross-parsing where a target could decide it. The six constant-returning
stubs get a call each again: the constant is this repo's own C, but the call
is what checks the binding still marshals.

CI, packaging and prose:

- deps/build_ci.jl decided whether to rebuild the shim by comparing the time
  of the last ClangExtra commit with the time of the JLL's tag, which says
  nothing about which LLVM that JLL was built for. One with no artifact for
  this platform and LLVM is now never "the JLL's support library", however
  recent its tag.
- The bindings job diffed `lib/` only, while the generator it runs also
  writes fourteen files under `src/`, and `git diff` does not see a new file.
  It fails on anything changed or added under either.
- The bindings and examples jobs get the timeout the test job has.
- deps/ClangExtra/CLAUDE.md describes the exhaustive-switch guard beside
  ENUM_SYNC, and what the bindings job now checks.
- MARSHALLING.md cited `Driver::getLTOMode(true)` and a line of a julia-18
  tag. LLVM 20 spells the offload mode `getOffloadLTOMode()`; the claim that
  only it is uninitialised still holds, and no longer carries a line number.
- examples/06 pointed at "§3" for the act that runs code; it is Act 4.
- The api-coverage skill named the LLVM 18 artifact by hash. It finds the
  LLVM_full artifact for the running Julia's LLVM by version, and no longer
  cites a test file that does not exist.

A hand-built `Module`'s bits. test/CLAUDE.md filed a synthetic module's
availability, its `isPartOfFramework` and the `markUnavailable` transition
under "nothing decides it": reads of memory nobody wrote, because such a module
never had a module map to set those bits. `clang::Module`'s constructor sets
every one of them in its mem-init list. They are equalities now and hold on all
three runners, Windows included, where an earlier disagreement had been taken
as proof of the opposite; `isPartOfFramework` is false for a plain module and
true for one constructed as a framework, the bit being a constructor argument.
The guidance says what the episode taught: "nothing wrote it" is a claim about
a mem-init list and is checked there, not inferred from how the object came to
exist. It also names the second way a value comes back outside its enum, which
this branch met ten times: a mirror that lacks an enumerator upstream appended.
`checkModuleIsAvailable`'s docstring called a hand-built module's availability
host-decided; a module is born available.

+++ deps/ClangExtra/include/clang-ex/AST/CXDecl.h
@@ -44,7 +44,7 @@ CXPragmaCommentDecl clang_PragmaCommentDecl_Create(CXASTContext C, CXTranslation
                                                    CXPragmaMSCommentKind CommentKind,
                                                    const char *Arg);
 
-CXPragmaCommentDecl clang_PragmaCommentDecl_CreateDeserialized(CXASTContext C, unsigned ID,
+CXPragmaCommentDecl clang_PragmaCommentDecl_CreateDeserialized(CXASTContext C, uint64_t ID,
                                                                unsigned ArgSize);
 
 CXPragmaMSCommentKind clang_PragmaCommentDecl_getCommentKind(CXPragmaCommentDecl PCD);
@@ -62,7 +62,7 @@ CXPragmaDetectMismatchDecl clang_PragmaDetectMismatchDecl_Create(CXASTContext C,
                                                                  const char *Value);
 
 CXPragmaDetectMismatchDecl
-clang_PragmaDetectMismatchDecl_CreateDeserialized(CXASTContext C, unsigned ID,
+clang_PragmaDetectMismatchDecl_CreateDeserialized(CXASTContext C, uint64_t ID,
                                                   unsigned NameValueSize);
 
 const char *clang_PragmaDetectMismatchDecl_getName(CXPragmaDetectMismatchDecl PDMD);
@@ -170,7 +170,7 @@ bool clang_NamedDecl_classofKind(CXDeclKind K);
 CXLabelDecl clang_LabelDecl_Create(CXASTContext C, CXDeclContext DC,
                                    CXSourceLocation_ IdentL, CXIdentifierInfo II);
 
-CXLabelDecl clang_LabelDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXLabelDecl clang_LabelDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXLabelStmt clang_LabelDecl_getStmt(CXLabelDecl LD);
 
@@ -210,7 +210,7 @@ CXNamespaceDecl clang_NamespaceDecl_Create(CXASTContext C, CXDeclContext DC, boo
                                            CXSourceLocation_ IdLoc, CXIdentifierInfo Id,
                                            CXNamespaceDecl PrevDecl, bool Nested);
 
-CXNamespaceDecl clang_NamespaceDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXNamespaceDecl clang_NamespaceDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 bool clang_NamespaceDecl_isAnonymousNamespace(CXNamespaceDecl ND);
 
@@ -344,7 +344,7 @@ CXVarDecl clang_VarDecl_Create(CXASTContext C, CXDeclContext DC, CXSourceLocatio
                                CXSourceLocation_ IdLoc, CXIdentifierInfo Id, CXQualType T,
                                CXTypeSourceInfo TInfo, CXStorageClass S);
 
-CXVarDecl clang_VarDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXVarDecl clang_VarDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXSourceRange_ clang_VarDecl_getSourceRange(CXVarDecl VD);
 
@@ -564,7 +564,7 @@ CXImplicitParamDecl clang_ImplicitParamDecl_Create(CXASTContext C, CXDeclContext
                                                    CXIdentifierInfo Id, CXQualType T,
                                                    CXImplicitParamKind ParamKind);
 
-CXImplicitParamDecl clang_ImplicitParamDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXImplicitParamDecl clang_ImplicitParamDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXImplicitParamKind clang_ImplicitParamDecl_getParameterKind(CXImplicitParamDecl IPD);
 
@@ -578,7 +578,7 @@ CXParmVarDecl clang_ParmVarDecl_Create(CXASTContext C, CXDeclContext DC,
                                        CXTypeSourceInfo TInfo, CXStorageClass S,
                                        CXExpr DefArg);
 
-CXParmVarDecl clang_ParmVarDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXParmVarDecl clang_ParmVarDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXSourceRange_ clang_ParmVarDecl_getSourceRange(CXParmVarDecl PVD);
 
@@ -687,7 +687,7 @@ CXFunctionDecl clang_FunctionDecl_Create(CXASTContext C, CXDeclContext DC,
                                          CXTypeSourceInfo TInfo, CXStorageClass SC,
                                          bool isInlineSpecified, bool hasWrittenPrototype);
 
-CXFunctionDecl clang_FunctionDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXFunctionDecl clang_FunctionDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 // Returns an owned box; release with clang_DeclarationNameInfo_dispose.
 CXDeclarationNameInfo clang_FunctionDecl_getNameInfo(CXFunctionDecl FD);
@@ -918,7 +918,8 @@ unsigned clang_FunctionDecl_getNumParams(CXFunctionDecl FD);
 
 CXParmVarDecl clang_FunctionDecl_getParamDecl(CXFunctionDecl FD, unsigned i);
 
-// setParams (private in clang 18 — only Sema/deserialization may set params)
+// setParams is not wrapped. Of FunctionDecl's two overloads only the one taking an ASTContext
+// is private; setParams(ArrayRef<ParmVarDecl *>) is public.
 
 unsigned clang_FunctionDecl_getMinRequiredArguments(CXFunctionDecl FD);
 
@@ -1059,7 +1060,7 @@ CXFieldDecl clang_FieldDecl_Create(CXASTContext C, CXDeclContext DC,
                                    CXIdentifierInfo I, CXQualType T, CXTypeSourceInfo TInfo,
                                    CXExpr BW, bool Mutable, CXInClassInitStyle InitStyle);
 
-CXFieldDecl clang_FieldDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXFieldDecl clang_FieldDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 unsigned clang_FieldDecl_getFieldIndex(CXFieldDecl FD);
 
@@ -1121,7 +1122,7 @@ CXEnumConstantDecl clang_EnumConstantDecl_Create(CXASTContext C, CXEnumDecl DC,
                                                  CXQualType T, CXExpr E,
                                                  LLVMGenericValueRef V);
 
-CXEnumConstantDecl clang_EnumConstantDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXEnumConstantDecl clang_EnumConstantDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXExpr clang_EnumConstantDecl_getInitExpr(CXEnumConstantDecl ECD);
 
@@ -1179,7 +1180,7 @@ CXIndirectFieldDecl clang_IndirectFieldDecl_Create(CXASTContext C, CXDeclContext
                                                    CXNamedDecl *Chain,
                                                    unsigned ChainSize);
 
-CXIndirectFieldDecl clang_IndirectFieldDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXIndirectFieldDecl clang_IndirectFieldDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 // chain
 CXNamedDecl clang_IndirectFieldDecl_getChainElement(CXIndirectFieldDecl IFD, unsigned i);
@@ -1238,7 +1239,7 @@ CXTypedefDecl clang_TypedefDecl_Create(CXASTContext C, CXDeclContext DC,
                                        CXSourceLocation_ StartLoc, CXSourceLocation_ IdLoc,
                                        CXIdentifierInfo Id, CXTypeSourceInfo TInfo);
 
-CXTypedefDecl clang_TypedefDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXTypedefDecl clang_TypedefDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXSourceRange_ clang_TypedefDecl_getSourceRange(CXTypedefDecl TD);
 
@@ -1251,7 +1252,7 @@ CXTypeAliasDecl clang_TypeAliasDecl_Create(CXASTContext C, CXDeclContext DC,
                                            CXSourceLocation_ IdLoc, CXIdentifierInfo Id,
                                            CXTypeSourceInfo TInfo);
 
-CXTypeAliasDecl clang_TypeAliasDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXTypeAliasDecl clang_TypeAliasDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXSourceRange_ clang_TypeAliasDecl_getSourceRange(CXTypeAliasDecl TAD);
 
@@ -1369,7 +1370,7 @@ CXEnumDecl clang_EnumDecl_Create(CXASTContext C, CXDeclContext DC,
                                  CXIdentifierInfo Id, CXEnumDecl PrevDecl, bool IsScoped,
                                  bool IsScopedUsingClassTag, bool IsFixed);
 
-CXEnumDecl clang_EnumDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXEnumDecl clang_EnumDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 void clang_EnumDecl_setScoped(CXEnumDecl ED, bool Scoped);
 
@@ -1465,7 +1466,7 @@ CXRecordDecl clang_RecordDecl_Create(CXASTContext C, CXTagTypeKind TK, CXDeclCon
                                      CXSourceLocation_ StartLoc, CXSourceLocation_ IdLoc,
                                      CXIdentifierInfo Id, CXRecordDecl PrevDecl);
 
-CXRecordDecl clang_RecordDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXRecordDecl clang_RecordDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXRecordDecl clang_RecordDecl_getPreviousDecl(CXRecordDecl RD);
 
@@ -1572,7 +1573,7 @@ CXFileScopeAsmDecl clang_FileScopeAsmDecl_Create(CXASTContext C, CXDeclContext D
                                                  CXSourceLocation_ AsmLoc,
                                                  CXSourceLocation_ RParenLoc);
 
-CXFileScopeAsmDecl clang_FileScopeAsmDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXFileScopeAsmDecl clang_FileScopeAsmDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXSourceLocation_ clang_FileScopeAsmDecl_getAsmLoc(CXFileScopeAsmDecl FSAD);
 
@@ -1592,7 +1593,7 @@ bool clang_FileScopeAsmDecl_classofKind(CXDeclKind K);
 // TopLevelStmtDecl
 CXTopLevelStmtDecl clang_TopLevelStmtDecl_Create(CXASTContext C, CXStmt Statement);
 
-CXTopLevelStmtDecl clang_TopLevelStmtDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXTopLevelStmtDecl clang_TopLevelStmtDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXSourceRange_ clang_TopLevelStmtDecl_getSourceRange(CXTopLevelStmtDecl TLSD);
 
@@ -1611,7 +1612,7 @@ bool clang_TopLevelStmtDecl_classofKind(CXDeclKind K);
 // BlockDecl
 CXBlockDecl clang_BlockDecl_Create(CXASTContext C, CXDeclContext DC, CXSourceLocation_ L);
 
-CXBlockDecl clang_BlockDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXBlockDecl clang_BlockDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXSourceLocation_ clang_BlockDecl_getCaretLocation(CXBlockDecl BD);
 
@@ -1721,7 +1722,7 @@ CXBlockDecl clang_BlockDecl_castFromDeclContext(CXDeclContext DC);
 CXCapturedDecl clang_CapturedDecl_Create(CXASTContext C, CXDeclContext DC,
                                          unsigned NumParams);
 
-CXCapturedDecl clang_CapturedDecl_CreateDeserialized(CXASTContext C, unsigned ID,
+CXCapturedDecl clang_CapturedDecl_CreateDeserialized(CXASTContext C, uint64_t ID,
                                                      unsigned NumParams);
 
 CXStmt clang_CapturedDecl_getBody(CXCapturedDecl CD);
@@ -1758,7 +1759,7 @@ CXImportDecl clang_ImportDecl_CreateImplicit(CXASTContext C, CXDeclContext DC,
                                              CXSourceLocation_ StartLoc, CXModule_ Imported,
                                              CXSourceLocation_ EndLoc);
 
-CXImportDecl clang_ImportDecl_CreateDeserialized(CXASTContext C, unsigned ID,
+CXImportDecl clang_ImportDecl_CreateDeserialized(CXASTContext C, uint64_t ID,
                                                  unsigned NumLocations);
 
 CXModule_ clang_ImportDecl_getImportedModule(CXImportDecl ID);
@@ -1777,7 +1778,7 @@ bool clang_ImportDecl_classofKind(CXDeclKind K);
 CXExportDecl clang_ExportDecl_Create(CXASTContext C, CXDeclContext DC,
                                      CXSourceLocation_ ExportLoc);
 
-CXExportDecl clang_ExportDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXExportDecl clang_ExportDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXSourceLocation_ clang_ExportDecl_getExportLoc(CXExportDecl ED);
 
@@ -1802,7 +1803,7 @@ CXExportDecl clang_ExportDecl_castFromDeclContext(CXDeclContext DC);
 // EmptyDecl
 CXEmptyDecl clang_EmptyDecl_Create(CXASTContext C, CXDeclContext DC, CXSourceLocation_ L);
 
-CXEmptyDecl clang_EmptyDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXEmptyDecl clang_EmptyDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 // The Decl::Kind test behind isa<EmptyDecl> (as above).
 bool clang_EmptyDecl_classofKind(CXDeclKind K);
@@ -1817,7 +1818,7 @@ CXHLSLBufferDecl clang_HLSLBufferDecl_Create(CXASTContext C, CXDeclContext Lexic
                                              CXIdentifierInfo ID, CXSourceLocation_ IDLoc,
                                              CXSourceLocation_ LBrace);
 
-CXHLSLBufferDecl clang_HLSLBufferDecl_CreateDeserialized(CXASTContext C, unsigned ID);
+CXHLSLBufferDecl clang_HLSLBufferDecl_CreateDeserialized(CXASTContext C, uint64_t ID);
 
 CXSourceRange_ clang_HLSLBufferDecl_getSourceRange(CXHLSLBufferDecl BD);
